adapting to the fact that getFloat -> getDouble in WPXProperty class

+++ filter/ListStyle.cxx
@@ -70,11 +70,11 @@ void OrderedListLevelStyle::write(DocumentHandler *pHandler, int iLevel) const
 	listLevelStyleOpen.write(pHandler);
 
 	TagOpenElement stylePropertiesOpen("style:list-level-properties");
-	if (mPropList["text:space-before"] && mPropList["text:space-before"]->getFloat() > 0.0f)
+	if (mPropList["text:space-before"] && mPropList["text:space-before"]->getDouble() > 0.0)
 		stylePropertiesOpen.addAttribute("text:space-before", mPropList["text:space-before"]->getStr());
-	if (mPropList["text:min-label-width"] && mPropList["text:min-label-width"]->getFloat() > 0.0f)
+	if (mPropList["text:min-label-width"] && mPropList["text:min-label-width"]->getDouble() > 0.0)
 		stylePropertiesOpen.addAttribute("text:min-label-width", mPropList["text:min-label-width"]->getStr());
-	if (mPropList["text:min-label-distance"] && mPropList["text:min-label-distance"]->getFloat() > 0.0f)
+	if (mPropList["text:min-label-distance"] && mPropList["text:min-label-distance"]->getDouble() > 0.0)
 		stylePropertiesOpen.addAttribute("text:min-label-distance", mPropList["text:min-label-distance"]->getStr());
 	stylePropertiesOpen.write(pHandler);
 
@@ -117,11 +117,11 @@ void UnorderedListLevelStyle::write(DocumentHandler *pHandler, int iLevel) const
 	listLevelStyleOpen.write(pHandler);
 
 	TagOpenElement stylePropertiesOpen("style:list-level-properties");
-	if (mPropList["text:space-before"] && mPropList["text:space-before"]->getFloat() > 0.0f)
+	if (mPropList["text:space-before"] && mPropList["text:space-before"]->getDouble() > 0.0)
 		stylePropertiesOpen.addAttribute("text:space-before", mPropList["text:space-before"]->getStr());
-	if (mPropList["text:min-label-width"] && mPropList["text:min-label-width"]->getFloat() > 0.0f)
+	if (mPropList["text:min-label-width"] && mPropList["text:min-label-width"]->getDouble() > 0.0)
 		stylePropertiesOpen.addAttribute("text:min-label-width", mPropList["text:min-label-width"]->getStr());
-	if (mPropList["text:min-label-distance"] && mPropList["text:min-label-distance"]->getFloat() > 0.0f)
+	if (mPropList["text:min-label-distance"] && mPropList["text:min-label-distance"]->getDouble() > 0.0)
 		stylePropertiesOpen.addAttribute("text:min-label-distance", mPropList["text:min-label-distance"]->getStr());
 	stylePropertiesOpen.addAttribute("style:font-name", "OpenSymbol");
 	stylePropertiesOpen.write(pHandler);
